# Angular-Cliente/client-management/src/app/swagger/model/client.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface Client { 
    sharedKey?: string;
    businessId?: string;
    email?: string;
    phone?: string;
    dateAdded?: string;
}